Updating Python debugger to use ms-python.debugpy & updating extension installation logic
